Fix legacy web app conflicts (#220)

+++ workflows/plugins/event_bus/swipe/event_bus_swipe.py
@@ -6,6 +6,7 @@ import json
 from typing import List, cast
 import boto3
 from settings import SWIPEEventBusSettings
+import uuid
 from plugins.plugin_types import (
     EventBus,
     WorkflowStartedMessage,
@@ -23,6 +24,13 @@ class EventBusSWIPE(EventBus):
         if settings.SQS_QUEUE_URL and settings.SQS_QUEUE_URL not in self.sqs.list_queues()["QueueUrls"]:
             raise Exception("SQS_QUEUE_URL not found")
 
+    def valid_uuid(self, execution_id: str) -> bool:
+        try:
+            uuid.UUID(execution_id, version=4)
+            return True
+        except ValueError:
+            return False
+
     def retrieve_messages(self, url: str) -> List:
         """Retrieve a list of SQS messages and delete them from queue"""
         resp = self.sqs.receive_message(
@@ -37,13 +45,16 @@ class EventBusSWIPE(EventBus):
         messages = []
         for message in resp["Messages"]:
             receipt_handle = message["ReceiptHandle"]
-            self.sqs.delete_message(
-                QueueUrl=url,
-                ReceiptHandle=receipt_handle,
-            )
             body = json.loads(message["Body"])
             content = json.loads(body["Message"]) if body.get("Message") else body
-            messages.append(content)
+
+            # check if message has valid uuid. If not could be from legacy web app
+            if self.valid_uuid(content["detail"]["executionArn"].split(":")[-1]):
+                messages.append(content)
+                self.sqs.delete_message(
+                    QueueUrl=url,
+                    ReceiptHandle=receipt_handle,
+                )
 
         return messages
 
@@ -96,8 +107,5 @@ class EventBusSWIPE(EventBus):
             elif message.get("source") == "aws.batch":
                 # TODO: return step status messages
                 pass
-
-            elif message.get("source") is None:
-                print("message missing source", message)
 
         return workflow_statuses
